perf(stdlib): optimize pivot for common scenarios
This optimizes pivot for the common scenario of a single row and single
column key where the column key is part of the group key.

It will perform a pivot by reading all of the relevant tables and it
will then merge all of the keys together in sorted order. After it has
done that, it will build each column by matching the value with the
appropriate key. This causes the algorithm to run in `O(n)` time and
avoid unnecessary reallocations because it determines the size before it
attempts to build any arrays.